Create simple_model
Simple scattering simulation based on the assumption that there are certain (randomly distributed) regions of increased electron density in the scattering screen which do all of the scattering. Their positions are chosen randomly at start and are defined by the angles in x and y direction from the source. The observer then moves parallel to the scattering screen with a given velocity, changing the angular positions of the scattering regions. At the detector, all arriving waves are interfered. This is repeated for a range of frequencies and the results are plotted as a dynamic spectrum.